chnaged grid to carousel
